refactor: address review — move schema tests, use ReplyToMode type
