+++ extensions/slack/src/monitor/channel-config.test.ts
@@ -1,4 +1,5 @@
 import { describe, expect, it } from "vitest";
+import { SlackChannelSchema } from "../../config/zod-schema.providers-core.js";
 import { resolveSlackChannelConfig } from "./channel-config.js";
 
 describe("resolveSlackChannelConfig replyToMode fields", () => {
@@ -35,5 +36,21 @@ describe("resolveSlackChannelConfig replyToMode fields", () => {
       channels: { C123: { allow: true } },
     });
     expect(result?.replyToMode).toBeUndefined();
+  });
+});
+
+describe("SlackChannelSchema replyToMode field", () => {
+  it("accepts valid replyToMode values", () => {
+    for (const value of ["off", "first", "all"] as const) {
+      const result = SlackChannelSchema.safeParse({ replyToMode: value });
+      expect(result.success, `expected "${value}" to be valid`).toBe(true);
+    }
+  });
+
+  it("rejects invalid replyToMode values", () => {
+    for (const value of ["invalid", true, 123]) {
+      const result = SlackChannelSchema.safeParse({ replyToMode: value });
+      expect(result.success, `expected ${JSON.stringify(value)} to be invalid`).toBe(false);
+    }
   });
 });
