+++ extensions/slack/src/monitor/channel-config.ts
@@ -5,6 +5,7 @@ import {
   type ChannelMatchSource,
 } from "../../../../src/channels/channel-config.js";
 import type { SlackReactionNotificationMode } from "../../../../src/config/config.js";
+import type { ReplyToMode } from "../../../../src/config/types.base.js";
 import type { SlackMessageEvent } from "../types.js";
 import { allowListMatches, normalizeAllowListLower, normalizeSlackSlug } from "./allow-list.js";
 
@@ -17,7 +18,7 @@ export type SlackChannelConfigResolved = {
   systemPrompt?: string;
   matchKey?: string;
   matchSource?: ChannelMatchSource;
-  replyToMode?: "off" | "first" | "all";
+  replyToMode?: ReplyToMode;
 };
 
 export type SlackChannelConfigEntry = {
@@ -28,7 +29,7 @@ export type SlackChannelConfigEntry = {
   users?: Array<string | number>;
   skills?: string[];
   systemPrompt?: string;
-  replyToMode?: "off" | "first" | "all";
+  replyToMode?: ReplyToMode;
 };
 
 export type SlackChannelConfigEntries = Record<string, SlackChannelConfigEntry>;
